+++ client/dashboard/profile-wizard/steps/product-types.js
@@ -14,6 +14,7 @@ import { withDispatch } from '@wordpress/data';
  * Internal dependencies
  */
 import { H, Card, Link } from '@woocommerce/components';
+import { getSetting } from '@woocommerce/wc-admin-settings';
 import withSelect from 'wc-api/with-select';
 import { recordEvent } from 'lib/tracks';
 
@@ -86,8 +87,8 @@ class ProductTypes extends Component {
 	}
 
 	render() {
-		const { productTypes } = wcSettings.onboarding;
-		const { error } = this.state;
+		const { productTypes = {} } = getSetting( 'onboarding', {} );
+		const { error, selected } = this.state;
 		return (
 			<Fragment>
 				<H className="woocommerce-profile-wizard__header-title">
@@ -134,6 +135,7 @@ class ProductTypes extends Component {
 						isPrimary
 						className="woocommerce-profile-wizard__continue"
 						onClick={ this.onContinue }
+						disabled={ ! selected.length }
 					>
 						{ __( 'Continue', 'woocommerce-admin' ) }
 					</Button>