+++ client/dashboard/task-list/tasks/payments/paypal.js
@@ -5,7 +5,7 @@
 import { __ } from '@wordpress/i18n';
 import { Component, Fragment } from '@wordpress/element';
 import apiFetch from '@wordpress/api-fetch';
-import { Button, TextControl } from 'newspack-components';
+import { Button } from 'newspack-components';
 import { getQuery } from '@woocommerce/navigation';
 import { withDispatch } from '@wordpress/data';
 import { compose } from '@wordpress/compose';
@@ -15,8 +15,9 @@ import interpolateComponents from 'interpolate-components';
  * WooCommerce dependencies
  */
 import { WC_ADMIN_NAMESPACE } from 'wc-api/constants';
-import { Form, Link } from '@woocommerce/components';
+import { Form, Link, TextControl } from '@woocommerce/components';
 import withSelect from 'wc-api/with-select';
+import { recordEvent } from 'lib/tracks';
 
 class PayPal extends Component {
 	constructor( props ) {
@@ -37,6 +38,7 @@ class PayPal extends Component {
 		// Handle redirect back from PayPal
 		if ( query[ 'paypal-connect' ] ) {
 			if ( '1' === query[ 'paypal-connect' ] ) {
+				recordEvent( 'tasklist_payment_connect_method', { payment_method: 'paypal' } );
 				this.props.markConfigured( 'paypal' );
 				this.props.createNotice(
 					'success',
@@ -124,6 +126,7 @@ class PayPal extends Component {
 		} );
 
 		if ( ! isSettingsError ) {
+			recordEvent( 'tasklist_payment_connect_method', { payment_method: 'paypal' } );
 			this.props.setRequestPending( false );
 			markConfigured( 'paypal' );
 			this.props.createNotice(