+++ packages/components/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 4.1.0 (Unreleased)
+- Renamed the `<Autocomplete />` component to `<SelectControl />`.
+- Added `isSearchable` prop to `<SelectControl />` to allow simple select dropdowns.
+- Removed the `<SimpleSelectControl />` component.
+- Removed WC-Admin specific actions from `<TableCard />` component.
+- Export the `<CompareButton />` component.
+- Add `<TextControl />` component.
+
 # 4.0.0
 - Added a new `<ScrollTo />` component.
 - Changed the `<List />` `description` prop to `content` and allowed content nodes to be passed in addition to strings.